ollama: merge chat into llm, implement Call with GenerateContent
